# docs/02-方案/接口契约_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "安联资管运维管理平台",
    "description": "安联资管运维管理平台（恒生 O32 投资交易系统日常运维）：M1/M2/M3 核对 + 数据源管理 + 系统配置 + 任务调度中心 + 统计看板 + 数据字典 + 本地认证",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dict/models": {
      "get": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "字典模型清单",
        "operationId": "list_models_api_dict_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DictModelInfo"
                  },
                  "type": "array",
                  "title": "Response List Models Api Dict Models Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/dict/tables": {
      "get": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "搜索表（表代码/中文名/字段名模糊，收藏优先置顶）",
        "operationId": "search_tables_api_dict_tables_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "default": "",
              "title": "Keyword"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictTableSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dict/tables/{table_id}": {
      "get": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "表详情（全字段 + 主键标识）",
        "operationId": "table_detail_api_dict_tables__table_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictTableDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dict/tables/{table_id}/references": {
      "get": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "表关联（父子两向）",
        "operationId": "table_references_api_dict_tables__table_id__references_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictTableReferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dict/favorites": {
      "get": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "当前用户收藏的字典表",
        "operationId": "list_favorites_api_dict_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DictFavoriteInfo"
                  },
                  "type": "array",
                  "title": "Response List Favorites Api Dict Favorites Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "收藏字典表",
        "operationId": "add_favorite_api_dict_favorites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictFavoriteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictFavoriteInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/dict/favorites/{table_id}": {
      "delete": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "取消收藏字典表",
        "operationId": "remove_favorite_api_dict_favorites__table_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dict/gen-sql": {
      "post": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "生成 Oracle SELECT（多表按外键自动 JOIN）",
        "operationId": "gen_sql_api_dict_gen_sql_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenSqlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenSqlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/dict/save-template": {
      "post": {
        "tags": [
          "数据字典查询"
        ],
        "summary": "生成的 SQL 保存为查询模板（模块=custom）",
        "operationId": "save_template_api_dict_save_template_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictSaveTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTemplateInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DictColumnInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "col_code": {
            "type": "string",
            "title": "Col Code"
          },
          "col_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Col Name"
          },
          "data_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Type"
          },
          "is_pk": {
            "type": "integer",
            "title": "Is Pk"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "col_code",
          "is_pk",
          "seq"
        ],
        "title": "DictColumnInfo"
      },
      "DictFavoriteCreateRequest": {
        "properties": {
          "table_id": {
            "type": "integer",
            "title": "Table Id",
            "description": "字典表 ID"
          },
          "table_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Table Code",
            "description": "表代码"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name",
            "description": "表中文名"
          }
        },
        "type": "object",
        "required": [
          "table_id",
          "table_code"
        ],
        "title": "DictFavoriteCreateRequest"
      },
      "DictFavoriteInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "table_id": {
            "type": "integer",
            "title": "Table Id"
          },
          "table_code": {
            "type": "string",
            "title": "Table Code"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "table_id",
          "table_code",
          "created_at"
        ],
        "title": "DictFavoriteInfo"
      },
...
      "DictModelInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "biz_group": {
            "type": "string",
            "title": "Biz Group"
          },
          "table_count": {
            "type": "integer",
            "title": "Table Count"
          },
          "column_count": {
            "type": "integer",
            "title": "Column Count"
          },
          "imported_at": {
            "type": "string",
            "title": "Imported At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_name",
          "model_name",
          "biz_group",
          "table_count",
          "column_count",
          "imported_at"
        ],
        "title": "DictModelInfo"
      },
...
      "DictSaveTemplateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "模板名称（唯一）"
          },
          "ds_id": {
            "type": "integer",
            "title": "Ds Id",
            "description": "数据源 ID"
          },
          "sql_text": {
            "type": "string",
            "minLength": 1,
            "title": "Sql Text",
            "description": "生成的 SQL（仅单条 SELECT/WITH）"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ds_id",
          "sql_text"
        ],
        "title": "DictSaveTemplateRequest"
      },
      "DictTableDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "model_id": {
            "type": "integer",
            "title": "Model Id"
          },
          "table_code": {
            "type": "string",
            "title": "Table Code"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "biz_group": {
            "type": "string",
            "title": "Biz Group"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/DictColumnInfo"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "id",
          "model_id",
          "table_code",
          "biz_group",
          "file_name",
          "model_name",
          "columns"
        ],
        "title": "DictTableDetail"
      },
      "DictTableItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "model_id": {
            "type": "integer",
            "title": "Model Id"
          },
          "table_code": {
            "type": "string",
            "title": "Table Code"
          },
          "table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Name"
          },
          "biz_group": {
            "type": "string",
            "title": "Biz Group"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "matched_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched On",
            "description": "命中点: table_code/table_name/column"
          },
          "matched_columns": {
            "items": {
              "$ref": "#/components/schemas/DictMatchedColumn"
            },
            "type": "array",
            "title": "Matched Columns"
          },
          "is_favorite": {
            "type": "boolean",
            "title": "Is Favorite",
            "description": "当前用户是否收藏",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "model_id",
          "table_code",
          "biz_group",
          "file_name"
        ],
        "title": "DictTableItem"
      },
      "DictTableReferencesResponse": {
        "properties": {
          "as_parent": {
            "items": {
              "$ref": "#/components/schemas/DictReferenceItem"
            },
            "type": "array",
            "title": "As Parent",
            "description": "本表为父表（被引用的子表列表）"
          },
          "as_child": {
            "items": {
              "$ref": "#/components/schemas/DictReferenceItem"
            },
            "type": "array",
            "title": "As Child",
            "description": "本表为子表（引用的父表列表）"
          }
        },
        "type": "object",
        "title": "DictTableReferencesResponse"
      },
      "DictTableSearchResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DictTableItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "DictTableSearchResponse"
      },
...
      "GenSqlRequest": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/GenSqlTableSpec"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Tables"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/GenSqlCondition"
            },
            "type": "array",
            "title": "Conditions"
          },
          "limit": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "ROWNUM 行数限制",
            "default": 500
          },
          "use_rownum": {
            "type": "boolean",
            "title": "Use Rownum",
            "description": "是否附加 ROWNUM 限制",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "tables"
        ],
        "title": "GenSqlRequest"
      },
      "GenSqlResponse": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "joins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Joins",
            "description": "自动 JOIN 说明"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "无关联等警告"
          }
        },
        "type": "object",
        "required": [
          "sql"
        ],
        "title": "GenSqlResponse"
      },
      "GenSqlTableSpec": {
        "properties": {
          "table_id": {
            "type": "integer",
            "title": "Table Id",
            "description": "字典表 ID"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns",
            "description": "勾选字段代码（空 = 全部字段）"
          },
          "alias": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias",
            "description": "表别名（默认 t1..tn）"
          }
        },
        "type": "object",
        "required": [
          "table_id"
        ],
        "title": "GenSqlTableSpec"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryTemplateInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "module": {
            "type": "string",
            "title": "Module"
          },
          "ds_id": {
            "type": "integer",
            "title": "Ds Id"
          },
          "ds_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ds Name"
          },
          "sql_text": {
            "type": "string",
            "title": "Sql Text"
          },
          "column_map": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Map"
          },
          "params_def": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params Def"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "module",
          "ds_id",
          "sql_text",
          "enabled",
          "updated_at"
        ],
        "title": "QueryTemplateInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT 访问令牌",
        "scheme": "bearer"
      }
    }
  }
}